长度使用 int
方便跟 len 等内置函数比较

+++ meta.go
@@ -11,7 +11,7 @@ type MetaResult struct {
 	Key        string
 	LastAccess uint64
 	Opaque     string
-	Size       uint64
+	Size       int
 	TTL        int64
 	Value      []byte
 	Hit        bool
